# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface NdRegisterInput {
  email: string;
  password: string;
  name?: string;
}

export interface NdLoginInput {
  email: string;
  password: string;
}

export interface NdAuthToken {
  token: string;
  accountId: number;
  email: string;
  /** @nullable */
  name?: string | null;
}

export interface NdApiKey {
  id: number;
  name: string;
  keyPrefix: string;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
}

export interface NdApiKeyInput {
  name: string;
}

export interface NdApiKeyCreated {
  id: number;
  name: string;
  keyPrefix: string;
  key: string;
  createdAt: string;
}

export interface NdProject {
  id: number;
  name: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  specs?: string | null;
  /** @nullable */
  drawingNotes?: string | null;
  status: string;
  /** @nullable */
  externalRef?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface NdProjectInput {
  name: string;
  industry?: string;
  specs?: string;
  drawingNotes?: string;
  externalRef?: string;
}

export interface NdProjectUpdate {
  status?: string;
  drawingNotes?: string;
  specs?: string;
}

export interface NdDrawing {
  id: number;
  projectId: number;
  fileName: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  uploadedBy?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface NdDrawingInput {
  fileName: string;
  fileUrl?: string;
  notes?: string;
}

export interface NdDashboardStats {
  totalProjects: number;
  pendingProjects: number;
  completeProjects: number;
  totalDrawings: number;
  totalKeys: number;
}

export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
  userId?: string;
}

export interface UserProfile {
  userId: string;
  aiName: string;
  aiPersonality: string;
  memories: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateUserProfileBody {
  aiName?: string;
  aiPersonality?: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}