# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /openai/profiles/{userId}:
    get:
      operationId: getOpenaiProfile
      tags: [openai]
      summary: Get a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateOpenaiProfile
      tags: [openai]
      summary: Update a user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOpenaiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateOpenaiImageResponse"
  /nd/auth/register:
    post:
      operationId: ndRegister
      tags: [nd]
      summary: Register a New Dimensions account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NdRegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdAuthToken"
        "400":
          description: Validation error
  /nd/auth/login:
    post:
      operationId: ndLogin
      tags: [nd]
      summary: Login and get a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NdLoginInput"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdAuthToken"
        "401":
          description: Invalid credentials
  /nd/keys:
    get:
      operationId: ndListKeys
      tags: [nd]
      summary: List API keys for the authenticated account
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NdApiKey"
    post:
      operationId: ndCreateKey
      tags: [nd]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NdApiKeyInput"
      responses:
        "201":
          description: Created — plaintext key returned once only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdApiKeyCreated"
  /nd/keys/{id}:
    delete:
      operationId: ndDeleteKey
      tags: [nd]
      summary: Delete an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /nd/projects:
    get:
      operationId: ndListProjects
      tags: [nd]
      summary: List all projects for the authenticated account
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NdProject"
    post:
      operationId: ndCreateProject
      tags: [nd]
      summary: Create a new CAD project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NdProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdProject"
  /nd/projects/{id}:
    get:
      operationId: ndGetProject
      tags: [nd]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdProject"
        "404":
          description: Not found
    patch:
      operationId: ndUpdateProject
      tags: [nd]
      summary: Update project status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NdProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdProject"
  /nd/projects/{id}/drawings:
    get:
      operationId: ndListDrawings
      tags: [nd]
      summary: List drawings attached to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drawings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NdDrawing"
    post:
      operationId: ndAddDrawing
      tags: [nd]
      summary: Attach a drawing to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NdDrawingInput"
      responses:
        "201":
          description: Drawing attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdDrawing"
  /nd/dashboard:
    get:
      operationId: ndDashboard
      tags: [nd]
      summary: Dashboard stats for the authenticated account
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NdDashboardStats"
components:
  schemas:
    NdRegisterInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
        name: { type: string }
    NdLoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    NdAuthToken:
      type: object
      required: [token, accountId, email]
      properties:
        token: { type: string }
        accountId: { type: integer }
        email: { type: string }
        name: { type: ["string", "null"] }
    NdApiKey:
      type: object
      required: [id, name, keyPrefix, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        keyPrefix: { type: string }
        lastUsedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    NdApiKeyInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
    NdApiKeyCreated:
      type: object
      required: [id, name, keyPrefix, key, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        keyPrefix: { type: string }
        key: { type: string }
        createdAt: { type: string }
    NdProject:
      type: object
      required: [id, name, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        industry: { type: ["string", "null"] }
        specs: { type: ["string", "null"] }
        drawingNotes: { type: ["string", "null"] }
        status: { type: string }
        externalRef: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    NdProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        industry: { type: string }
        specs: { type: string }
        drawingNotes: { type: string }
        externalRef: { type: string }
    NdProjectUpdate:
      type: object
      properties:
        status: { type: string }
        drawingNotes: { type: string }
        specs: { type: string }
    NdDrawing:
      type: object
      required: [id, projectId, fileName, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        fileName: { type: string }
        fileUrl: { type: ["string", "null"] }
        uploadedBy: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    NdDrawingInput:
      type: object
      required: [fileName]
      properties:
        fileName: { type: string }
        fileUrl: { type: string }
        notes: { type: string }
    NdDashboardStats:
      type: object
      required: [totalProjects, pendingProjects, completeProjects, totalDrawings, totalKeys]
      properties:
        totalProjects: { type: integer }
        pendingProjects: { type: integer }
        completeProjects: { type: integer }
        totalDrawings: { type: integer }
        totalKeys: { type: integer }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
        userId:
          type: string
      required:
        - content
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        aiName:
          type: string
        aiPersonality:
          type: string
        memories:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - userId
        - aiName
        - aiPersonality
        - memories
        - createdAt
        - updatedAt
    UpdateUserProfileBody:
      type: object
      properties:
        aiName:
          type: string
        aiPersonality:
          type: string
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    GenerateOpenaiImageBody:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "512x512", "256x256"]
      required:
        - prompt
    GenerateOpenaiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error