# lib/api-zod/src/generated/types/ndApiKey.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NdApiKey {
  id: number;
  name: string;
  keyPrefix: string;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
}