# lib/api-zod/src/generated/types/ndDrawing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NdDrawing {
  id: number;
  projectId: number;
  fileName: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  uploadedBy?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}