# lib/api-zod/src/generated/types/ndProject.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NdProject {
  id: number;
  name: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  specs?: string | null;
  /** @nullable */
  drawingNotes?: string | null;
  status: string;
  /** @nullable */
  externalRef?: string | null;
  createdAt: string;
  updatedAt: string;
}