# lib/api-zod/src/generated/types/ndProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NdProjectInput {
  name: string;
  industry?: string;
  specs?: string;
  drawingNotes?: string;
  externalRef?: string;
}